falsifeye/requirements.txt
flask
reportlab
Pillow
textblob
opencv-python
librosa
numpy
scipy
PyPDF2
openai-whisper
sentence-transformers
nltk
torch
scikit-learn
numba
